March 16, 2016
March 16, 2016

+++ migrations/20151211182641-create_vendor_contact_address_book.js
@@ -37,11 +37,11 @@ module.exports = {
         allowNull:true
       },
       latitude:{
-        type:Sequelize.DECIMAL(10, 2),
+        type:Sequelize.FLOAT,
         allowNull:true
       },
       longitude:{
-        type:Sequelize.DECIMAL(10, 2),
+        type:Sequelize.FLOAT,
         allowNull:true
       },
       email:{
